split weak-ssl in 2 rules which are then merged
